# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "JWT Login API",
        "contact": {
            "name": "Junio Cesar Ferreira"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0.0"
    },
    "basePath": "/",
    "paths": {
        "/login": {
            "post": {
                "description": "Verifica se usuário é registrado e realiza login",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Login"
                ],
                "summary": "Tenta validar usuário e fazer login",
                "parameters": [
                    {
                        "description": "Usuário e Hash",
                        "name": "LoginRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.AuthData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/protected": {
            "get": {
                "description": "Testando Token",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Login"
                ],
                "summary": "Rota protegida com autenticação utilizando Token",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/renew": {
            "get": {
                "description": "Renova o token de autenticação se o token atual for válido e está prestes a expirar",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Login"
                ],
                "summary": "Renova o token de autenticação",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Token renovado com sucesso",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "auth.AuthData": {
            "type": "object",
            "properties": {
                "pass": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                }
            }
        }
    }
}